LayoutTests/media/modern-media-controls/start-support/start-support-fullscreen.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">
    
    video, #host {
        position: absolute;
        top: 0;
        left: 0;
        width: 320px;
        height: 240px;
    }
    
</style>
<video src="../../content/test.mp4"></video>
<div id="host"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>StartSupport</code> behavior with no source.");

const host = document.querySelector("div#host");
const media = document.querySelector("video");
const mediaController = createControls(host, media, null);

const button = document.body.appendChild(document.createElement("div"));
button.textContent = "Enter Fullscreen";
button.addEventListener("click", event => {
    try {
        media.webkitEnterFullscreen();
    } catch(e) {
        debug("Calling media.webkitEnterFullscreen() failed");
        finishJSTest();
    }
});

media.addEventListener("webkitfullscreenchange", function() {
    debug("Media entered fullscreen");
    shouldBeFalse("mediaController.controls.showsStartButton");
    debug("");
    host.remove();
    media.remove();
    button.remove();
    finishJSTest();
});

media.addEventListener("loadedmetadata", event => pressOnElement(button));

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
